+++ src/Data/Extensible/Field.hs
@@ -49,6 +49,7 @@ import Data.Coerce
 #if __GLASGOW_HASKELL__ < 802
 import Data.Constraint
 #endif
+import qualified Data.Csv as Csv
 import Data.Extensible.Class
 import Data.Extensible.Sum
 import Data.Extensible.Match
@@ -121,6 +122,7 @@ ND_Field(Bounded)
 ND_Field(NFData)
 ND_Field(Arbitrary)
 ND_Field(Hashable)
+ND_Field(Csv.FromField)
 
 newtype instance U.MVector s (Field h x) = MV_Field (U.MVector s (h (AssocValue x)))
 newtype instance U.Vector (Field h x) = V_Field (U.Vector (h (AssocValue x)))